refactor: move jenkins related function to jenkins package
