+++ pkg/jenkins/utils.go
@@ -183,3 +183,20 @@ func loadChildJobs(jenkinsClient gojenkins.JenkinsClient, name string) ([]*gojen
 	}
 	return answer, nil
 }
+
+// Jobname returns the Jenkins job name starting with the given prefix
+func JobName(prefix string, j *gojenkins.Job) string {
+	name := j.FullName
+	if name == "" {
+		name = j.Name
+	}
+	if prefix != "" {
+		name = prefix + "/" + name
+	}
+	return name
+}
+
+// IsPipeline checks if the job is a pipeline job
+func IsPipeline(j *gojenkins.Job) bool {
+	return strings.Contains(j.Class, "Job")
+}
